pyproject.toml
[tool.poetry]
name = "store"
version = "0.1.0"
description = ""
authors = ["vicjan89 <vicjan89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
